+++ src/main/java/com/acs/gameeng/base/Vector3D.java
@@ -1,15 +1,15 @@
 package com.acs.gameeng.base;
 
 public class Vector3D {
-    public float x;
-    public float y;
-    public float z;
-    public float w = 1;
+    public double x;
+    public double y;
+    public double z;
+    public double w = 1;
 
     public Vector3D() {
     }
 
-    public Vector3D(float x, float y, float z) {
+    public Vector3D(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
